+++ gojmx/exports_test.go
@@ -1,12 +1,12 @@
-package nrprotocol
+package gojmx
 
 import (
-	"testing"
-
+	"github.com/newrelic/nrjmx/gojmx/internal/nrprotocol"
 	"github.com/stretchr/testify/assert"
+	"testing"
 )
 
-func Test_Convert_Success(t *testing.T) {
+func Test_JMXAttribute_GetValue(t *testing.T) {
 
 	testCases := []struct {
 		name     string
@@ -17,16 +17,16 @@ func Test_Convert_Success(t *testing.T) {
 			name: "Double Value",
 			jmxAttr: &JMXAttribute{
 				Attribute:   "test:type=Cat,name=tomas,attr=FloatValue",
-				ValueType:   ValueType_DOUBLE,
+				ValueType:   nrprotocol.ValueType_DOUBLE,
 				DoubleValue: 2.222222,
 			},
-			expected: float64(2.222222),
+			expected: 2.222222,
 		},
 		{
 			name: "Number Value",
 			jmxAttr: &JMXAttribute{
 				Attribute: "test:type=Cat,name=tomas,attr=NumberValue",
-				ValueType: ValueType_INT,
+				ValueType: nrprotocol.ValueType_INT,
 				IntValue:  3,
 			},
 			expected: int64(3),
@@ -35,28 +35,28 @@ func Test_Convert_Success(t *testing.T) {
 			name: "Bool Value",
 			jmxAttr: &JMXAttribute{
 				Attribute: "test:type=Cat,name=tomas,attr=BoolValue",
-				ValueType: ValueType_BOOL,
+				ValueType: nrprotocol.ValueType_BOOL,
 				BoolValue: true,
 			},
-			expected: bool(true),
+			expected: true,
 		},
 		{
 			name: "Double Value",
 			jmxAttr: &JMXAttribute{
 				Attribute:   "test:type=Cat,name=tomas,attr=DoubleValue",
-				ValueType:   ValueType_DOUBLE,
+				ValueType:   nrprotocol.ValueType_DOUBLE,
 				DoubleValue: 1.2,
 			},
-			expected: float64(1.2),
+			expected: 1.2,
 		},
 		{
 			name: "String Value",
 			jmxAttr: &JMXAttribute{
 				Attribute:   "test:type=Cat,name=tomas,attr=Name",
-				ValueType:   ValueType_STRING,
+				ValueType:   nrprotocol.ValueType_STRING,
 				StringValue: "tomas",
 			},
-			expected: string("tomas"),
+			expected: "tomas",
 		},
 	}
